refactor(utils): add getDiffInCalendarMinutes as ESM
