+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as getDiffInMilliseconds } from './getDiffInMilliseconds.js';
 export { default as getDiffInCalendarSeconds } from './getDiffInCalendarSeconds.js';
+export { default as getDiffInCalendarMinutes } from './getDiffInCalendarMinutes.js';
